MySQL changes
Replaced `mysql` with `@mysql/xdevapi`. This allows for NoSQL operations

Created TypeScript type definitions for the library. Also created a more broad
	version to be pushed to GitHub (the one here is more strict - I like more
	strict)

+++ server/src/api/files/files/deletefile.ts
@@ -4,26 +4,48 @@ import { join } from 'path';
 import { Configuration } from '../../../conf';
 import { AccountRequest } from '../../../lib/Account';
 import { MemberRequest } from '../../../lib/BaseMember';
+import { collectResults } from '../../../lib/MySQLUtil';
 
 export default async (req: AccountRequest & MemberRequest, res: express.Response) => {
 	if (
 		typeof req.account !== 'undefined'
 	) {
-		await req.connectionPool.query(
-			'DELETE FROM FileInfo WHERE id = ? AND AccountID = ?',
-			[req.params.fileid, req.account.id],
+		const filesCollection = req.mysqlx.getCollection<FileObject>('Files');
+
+		const currentFiles = await collectResults(
+			filesCollection
+				.find('accountID = :accountID AND id = :id')
+				.bind({
+					accountID: req.account.id,
+					id: req.params.id
+				})
 		);
 
-		fs.unlink(join(Configuration.fileStoragePath, req.params.id), err => {
-			if (err) {
-				res.status(500);
-				res.end();
-			} else {
-				res.status(204);
-				res.end();
-			}
-		});
+		if (currentFiles.length !== 1) {
+			res.send(400);
+			return;
+		}
 
+		await filesCollection
+			.remove('accountID = :accountID AND id = :id')
+			.bind({
+				accountID: req.account.id,
+				id: req.params.id
+			})
+			.execute();
+
+		if (currentFiles[0].contentType !== 'application/folder') {
+			// These files don't get an associated hard disk file
+			fs.unlink(join(Configuration.fileStoragePath, req.params.id), err => {
+				if (err) {
+					res.status(500);
+					res.end();
+				} else {
+					res.status(204);
+					res.end();
+				}
+			});
+		}
 	} else {
 		res.status(400);
 		res.end();
